[Core][Frontend] Restructure gradient feature for less intrusive integration
Extract gradient-specific logic from core engine files into dedicated
modules, reducing cognitive load for developers working on inference:

- Move GradientOutput and GradientRequestOutput from outputs.py into
  vllm/entrypoints/gradient/outputs.py with re-exports for backward
  compatibility (-64 lines from outputs.py)

- Extract gradient queue/handler from core.py into a composed
  GradientHandler class in vllm/v1/engine/gradient_handler.py
  (-85 lines from core.py)

- Move numpy deserialization from output_processor.py into a factory
  function in the gradient outputs module (-35 lines from
  output_processor.py)

Net: -184 lines removed from core files. Same runtime behavior,
same IPC format, same execution order.
