+++ vllm/v1/engine/output_processor.py
@@ -17,7 +17,6 @@ from vllm.model_executor.layers.fused_moe.routed_experts_capturer import (
 from vllm.outputs import (
     STREAM_FINISHED,
     CompletionOutput,
-    GradientOutput,
     GradientRequestOutput,
     PoolingOutput,
     PoolingRequestOutput,
@@ -477,38 +476,16 @@ class RequestState:
         gradient_output: dict[str, Any],
         finished: bool,
     ) -> GradientRequestOutput:
+        from vllm.entrypoints.gradient.outputs import (
+            gradient_request_output_from_engine_dict,
+        )
+
         prompt_token_ids = self.prompt_token_ids
         if prompt_token_ids is None:
             prompt_token_ids = list(range(self.prompt_len))
 
-        # Reconstruct numpy arrays from packed bytes format.
-        token_attributions = None
-        if "token_attributions_bytes" in gradient_output:
-            token_attributions = np.frombuffer(
-                gradient_output["token_attributions_bytes"],
-                dtype=np.dtype(gradient_output["token_attributions_dtype"]),
-            ).reshape(gradient_output["token_attributions_shape"])
-
-        loss_gradients = None
-        if "loss_gradients_packed" in gradient_output:
-            loss_gradients = {}
-            for k, packed in gradient_output["loss_gradients_packed"].items():
-                loss_gradients[k] = np.frombuffer(
-                    packed["bytes"],
-                    dtype=np.dtype(packed["dtype"]),
-                ).reshape(packed["shape"])
-
-        return GradientRequestOutput(
-            request_id=external_req_id,
-            outputs=GradientOutput(
-                token_log_probs=gradient_output.get("token_log_probs"),
-                token_attributions=token_attributions,
-                loss=gradient_output.get("loss"),
-                loss_gradients=loss_gradients,
-            ),
-            prompt_token_ids=prompt_token_ids,
-            target_token_ids=gradient_output.get("target_token_ids", []),
-            finished=finished,
+        return gradient_request_output_from_engine_dict(
+            external_req_id, gradient_output, prompt_token_ids, finished
         )
